requirements.txt
coqui-tts
nltk
psutil
pydub
tqdm
ocrmypdf
pyspellchecker
torch
opencv-python
pdf2image
numpy
scipy
pdftotext
sentencepiece
ebooklib
BeautifulSoup4
streamlit
cython
